TrueCrypt v7.1 - TrueCrypt 7.1 Source.zip
extracted to src/

gpg verify TrueCrypt 7.1 Source.zip.sig
gpg: Signature made Thu, Sep 01, 2011 11:50:53 EDT using DSA key ID F0D6B1E0
gpg: Good signature from "TrueCrypt Foundation <info@truecrypt-foundation.org>"
gpg: WARNING: This key is not certified with a trusted signature!
gpg:          There is no indication that the signature belongs to the owner.
Primary key fingerprint: C5F4 BAC4 A7B2 2DB8 B8F8  5538 E3BA 73CA F0D6 B1E0

import script note: step 46 formula 20 common 16

+++ src/Setup/SelfExtract.c
@@ -537,11 +537,11 @@ static void FreeAllFileBuffers (void)
 // For details, see the definition of the DECOMPRESSED_FILE structure.
 BOOL SelfExtractInMemory (char *path)
 {
-	unsigned int filePos = 0, fileNo = 0;
-	unsigned int fileDataEndPos = 0;
-	unsigned int fileDataStartPos = 0;
-	unsigned int uncompressedLen = 0;
-	unsigned int compressedLen = 0;
+	int filePos = 0, fileNo = 0;
+	int fileDataEndPos = 0;
+	int fileDataStartPos = 0;
+	int uncompressedLen = 0;
+	int compressedLen = 0;
 	unsigned char *compressedData = NULL;
 	unsigned char *bufPos = NULL, *bufEndPos = NULL;
 
@@ -612,7 +612,7 @@ BOOL SelfExtractInMemory (char *path)
 	}
 
 	// Decompress the data
-	if ((unsigned int) DecompressBuffer (DecompressedData, compressedData, compressedLen) != uncompressedLen)
+	if (DecompressBuffer (DecompressedData, compressedData, compressedLen) != uncompressedLen)
 	{
 		Error ("DIST_PACKAGE_CORRUPTED");
 		goto sem_end;
